+++ lib/custom_refresh_view.dart
@@ -1,0 +1,3 @@
+library custom_refresh_view;
+
+export 'src/custom_refresh_view.dart';