+++ lib/src/custom_refresh_view.dart
@@ -1,0 +1,38 @@
+import 'package:easy_refresh/easy_refresh.dart';
+import 'package:flutter/material.dart';
+
+class CustomRefreshView extends StatelessWidget {
+  const CustomRefreshView({
+    super.key,
+    required this.itemCount,
+    required this.itemBuilder,
+    this.onRefresh,
+    this.onLoad,
+    this.padding = EdgeInsetsDirectional.zero,
+    this.controller,
+    this.physics,
+  });
+
+  final int itemCount;
+  final IndexedWidgetBuilder itemBuilder;
+  final Future<void> Function()? onRefresh;
+  final Future<void> Function()? onLoad;
+  final EdgeInsetsGeometry padding;
+  final ScrollController? controller;
+  final ScrollPhysics? physics;
+
+  @override
+  Widget build(BuildContext context) {
+    return EasyRefresh(
+      onRefresh: onRefresh,
+      onLoad: onLoad,
+      child: ListView.builder(
+        controller: controller,
+        physics: physics ?? const AlwaysScrollableScrollPhysics(),
+        padding: padding,
+        itemCount: itemCount,
+        itemBuilder: itemBuilder,
+      ),
+    );
+  }
+}