+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+### Added
+
+- Initial release of `custom_refresh_view`
+- `CustomRefreshView` with pull-to-refresh and load-more callbacks
+- LTR and RTL layout support